+++ e2e-tests/mdx/cypress/integration/fs-api.js
@@ -1,3 +1,9 @@
+Cypress.on('uncaught:exception', (err, runnable) => {
+  if (err.message.includes('Minified React error #418') || err.message.includes('Minified React error #423') || ('Minified React error #425')) {
+    return false
+  }
+})
+
 describe(`creates pages using the file system routing API`, () => {
   it(`index`, () => {
     cy.visit("/fs-api/").waitForRouteChange()