+++ e2e-tests/production-runtime/cypress/integration/lifecycle-methods.js
@@ -1,3 +1,9 @@
+Cypress.on('uncaught:exception', (err) => {
+  if ((err.message.includes('Minified React error #418') || err.message.includes('Minified React error #423') || err.message.includes('Minified React error #425')) && Cypress.env(`TEST_PLUGIN_OFFLINE`)) {
+    return false
+  }
+})
+
 describe(`Production build tests`, () => {
   it(`should remount when navigating to different template`, () => {
     cy.visit(`/`).waitForRouteChange()
@@ -9,7 +15,7 @@ describe(`Production build tests`, () => {
 
     // we expect 2 `componentDidMount` calls - 1 for initial page and 1 for second page
     cy.lifecycleCallCount(`componentDidMount`).should(`equal`, 2)
-    cy.lifecycleCallCount(`render`).should(`equal`, 2)
+    cy.lifecycleCallCount(`render`).should(`equal`, Cypress.env(`TEST_PLUGIN_OFFLINE`) ? 3 : 2)
   })
 
   it(`should remount when navigating to different page using same template`, () => {
@@ -22,7 +28,7 @@ describe(`Production build tests`, () => {
 
     // we expect 2 `componentDidMount` calls - 1 for initial page and 1 for duplicated page
     cy.lifecycleCallCount(`componentDidMount`).should(`equal`, 2)
-    cy.lifecycleCallCount(`render`).should(`equal`, 2)
+    cy.lifecycleCallCount(`render`).should(`equal`, Cypress.env(`TEST_PLUGIN_OFFLINE`) ? 3 : 2)
   })
 
   it(`should NOT remount when navigating within client only paths`, () => {
@@ -37,6 +43,6 @@ describe(`Production build tests`, () => {
 
     // we expect just 1 `componentDidMount` call, when navigating inside matchPath
     cy.lifecycleCallCount(`componentDidMount`).should(`equal`, 1)
-    cy.lifecycleCallCount(`render`).should(`equal`, 3)
+    cy.lifecycleCallCount(`render`).should(`equal`, Cypress.env(`TEST_PLUGIN_OFFLINE`) ? 4 : 3)
   })
 })