+++ examples/ecommerce-tutorial-with-stripe/src/components/Products/Products.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { graphql, StaticQuery } from "gatsby"
+import { graphql, useStaticQuery } from "gatsby"
 import ProductCard from "./ProductCard"
 
 const containerStyles = {
@@ -11,50 +11,45 @@ const containerStyles = {
 }
 
 const Products = () => {
-  return (
-    <StaticQuery
-      query={graphql`
-        query ProductPrices {
-          prices: allStripePrice(
-            filter: { active: { eq: true } }
-            sort: { fields: [unit_amount] }
-          ) {
-            edges {
-              node {
-                id
-                active
-                currency
-                unit_amount
-                product {
-                  id
-                  name
-                }
-              }
+  const { prices } = useStaticQuery(graphql`
+    query ProductPrices {
+      prices: allStripePrice(
+        filter: { active: { eq: true } }
+        sort: { unit_amount: ASC }
+      ) {
+        edges {
+          node {
+            id
+            active
+            currency
+            unit_amount
+            product {
+              id
+              name
             }
           }
         }
-      `}
-      render={({ prices }) => {
-        // Group prices by product
-        const products = {}
-        for (const { node: price } of prices.edges) {
-          const product = price.product
-          if (!products[product.id]) {
-            products[product.id] = product
-            products[product.id].prices = []
-          }
-          products[product.id].prices.push(price)
-        }
+      }
+    }
+  `)
 
-        return (
-          <div style={containerStyles}>
-            {Object.keys(products).map(key => (
-              <ProductCard key={products[key].id} product={products[key]} />
-            ))}
-          </div>
-        )
-      }}
-    />
+  // Group prices by product
+  const products = {}
+  for (const { node: price } of prices.edges) {
+    const product = price.product
+    if (!products[product.id]) {
+      products[product.id] = product
+      products[product.id].prices = []
+    }
+    products[product.id].prices.push(price)
+  }
+
+  return (
+    <div style={containerStyles}>
+      {Object.keys(products).map(key => (
+        <ProductCard key={products[key].id} product={products[key]} />
+      ))}
+    </div>
   )
 }