+++ packages/gatsby/src/internal-plugins/functions/middleware.ts
@@ -1,4 +1,4 @@
-import { match as reachMatch } from "@gatsbyjs/reach-router/lib/utils"
+import { match as reachMatch } from "@gatsbyjs/reach-router"
 import cookie from "cookie"
 import { urlencoded, text, json, raw } from "express"
 import type { RequestHandler, Request, Response, NextFunction } from "express"