+++ packages/gatsby/src/query/__tests__/data-tracking.js
@@ -281,6 +281,8 @@ const setup = async ({ restart = isFirstRun, clearCache = false } = {}) => {
   }
 }
 
+const indexBarFooArray = [`/`, `/bar/`, `/foo/`]
+
 describe(`query caching between builds`, () => {
   beforeAll(() => {
     setAPIhooks({
@@ -376,10 +378,10 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`, `/foo`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual(indexBarFooArray)
       expect(staticQueriesThatRan).toEqual([`static-query-1`])
     }, 99999)
 
@@ -388,7 +390,7 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // no changes should mean no queries to run
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -402,7 +404,7 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // no changes should mean no queries to run
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -417,10 +419,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`, `/foo`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual(indexBarFooArray)
       expect(staticQueriesThatRan).toEqual([`static-query-1`])
     }, 99999)
   })
@@ -466,10 +468,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`, `/foo`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual(indexBarFooArray)
       expect(staticQueriesThatRan).toEqual([`static-query-1`])
     }, 99999)
 
@@ -478,7 +480,7 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // it should not rerun any page query
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -492,7 +494,7 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // it should not rerun any page query
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -541,10 +543,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`, `/foo`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual(indexBarFooArray)
       expect(staticQueriesThatRan).toEqual([`static-query-1`])
     }, 99999)
 
@@ -553,10 +555,10 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
-      // we changed one node, so connection (`/`) and one detail page (`/bar`) should run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`])
+      // we changed one node, so connection (`/`) and one detail page (`/bar/`) should run
+      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar/`])
       expect(staticQueriesThatRan).toEqual([])
     }, 999999)
 
@@ -567,10 +569,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
-      // we changed one node, so connection (`/`) and one detail page (`/bar`) should run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`])
+      // we changed one node, so connection (`/`) and one detail page (`/bar/`) should run
+      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar/`])
       expect(staticQueriesThatRan).toEqual([])
     }, 999999)
   })
@@ -619,10 +621,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/`, `/bar`, `/foo`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual(indexBarFooArray)
       expect(staticQueriesThatRan).toEqual([`static-query-1`])
     }, 99999)
 
@@ -631,7 +633,7 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // no queries should run
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -645,7 +647,7 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/`, `/bar`, `/foo`])
+      expect(pages).toEqual(indexBarFooArray)
 
       // no queries should run
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -1257,10 +1259,10 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/no-dep-page`])
+      expect(pages).toEqual([`/no-dep-page/`])
 
       // on initial we want all queries to run
-      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/no-dep-page`])
+      expect(pathsOfPagesWithQueriesThatRan).toEqual([`/no-dep-page/`])
       expect(staticQueriesThatRan).toEqual([])
     }, 99999)
 
@@ -1269,7 +1271,7 @@ describe(`query caching between builds`, () => {
         await setup()
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/no-dep-page`])
+      expect(pages).toEqual([`/no-dep-page/`])
 
       // no queries should run
       expect(pathsOfPagesWithQueriesThatRan).toEqual([])
@@ -1283,7 +1285,7 @@ describe(`query caching between builds`, () => {
         })
 
       // sanity check, to make sure test setup is correct
-      expect(pages).toEqual([`/no-dep-page`])
+      expect(pages).toEqual([`/no-dep-page/`])
 
       // no queries should run
       // Currently it actually re-run query for `/no-dep-page`
@@ -1338,8 +1340,12 @@ describe(`query caching between builds`, () => {
                 },
                 barList: {
                   type: [`Bar`],
-                  resolve: (value, args, context) =>
-                    context.nodeModel.getAllNodes({ type: `Bar` }),
+                  resolve: async (value, args, context) => {
+                    const { entries } = await context.nodeModel.findAll({
+                      type: `Bar`,
+                    })
+                    return entries
+                  },
                 },
               },
               interfaces: [`Node`],
@@ -1447,14 +1453,14 @@ describe(`query caching between builds`, () => {
           clearCache: true,
         })
         // sanity check, to make sure test setup is correct
-        expect(pages).toEqual([`/bar`, `/bar-list`, `/foo`, `/foo-list`])
+        expect(pages).toEqual([`/bar-list/`, `/bar/`, `/foo-list/`, `/foo/`])
 
         // on initial we want query to run
         expect(pathsOfPagesWithQueriesThatRan).toEqual([
-          `/bar`,
-          `/bar-list`,
-          `/foo`,
-          `/foo-list`,
+          `/bar-list/`,
+          `/bar/`,
+          `/foo-list/`,
+          `/foo/`,
         ])
       }, 999999)
 
@@ -1466,7 +1472,7 @@ describe(`query caching between builds`, () => {
 
         // Since there was no single "foo" node before - we will also re-run the "/foo"
         // (as maybe the new node matches this query now)
-        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/foo`, `/foo-list`])
+        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/foo-list/`, `/foo/`])
       }, 999999)
 
       it(`re-runs queries when a foo node is deleted`, async () => {
@@ -1475,7 +1481,7 @@ describe(`query caching between builds`, () => {
           restart: withRestarts,
         })
 
-        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/foo`, `/foo-list`])
+        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/foo-list/`, `/foo/`])
       }, 999999)
 
       it(`re-runs queries when a bar-2 node is added`, async () => {
@@ -1489,7 +1495,7 @@ describe(`query caching between builds`, () => {
         // FIXME: this can be actually wrong behavior when filtering by any field other than "id"
         //   we should rework this to use connection + filter for tracking.
         //   In this case this test will also re-run "/bar" query (also affects other bar-related tests)
-        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar-list`])
+        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar-list/`])
       }, 999999)
 
       it(`re-runs queries when a bar-2 node is deleted`, async () => {
@@ -1498,7 +1504,7 @@ describe(`query caching between builds`, () => {
           restart: withRestarts,
         })
 
-        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar-list`])
+        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar-list/`])
       }, 999999)
 
       it(`re-runs queries when a bar-1 node is deleted`, async () => {
@@ -1507,7 +1513,7 @@ describe(`query caching between builds`, () => {
           restart: withRestarts,
         })
 
-        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar`, `/bar-list`])
+        expect(pathsOfPagesWithQueriesThatRan).toEqual([`/bar-list/`, `/bar/`])
       }, 999999)
     }